fixed the drift
fixed the drift the using the worldUp instead of the cameraUp for deriving the rightVector

+++ engine/src/engine.cpp
@@ -57,7 +57,7 @@ namespace braque
       // update the camera
       // update the input based on time left
       auto latency = swapchain.getFrameStats().Latency();
-      latency = std::min<double>(latency, max_latency);
+      latency = std::min<float>(latency, max_latency);
 
       accumulatedTime += latency;
